Move registered span types into a separate file
